examples/llmops/finetuning-flan-t5/AutoMLOps/components/component_base/requirements.txt
datasets==2.9.0
evaluate==0.4.0
google-cloud-aiplatform==1.26.0
google-cloud-storage==2.7.0
kfp<2.0.0
nltk==3.7
py7zr==0.20.4
rouge_score==0.1.2
tensorboard==2.11.2
transformers==4.25.1
